Dimensional reduction plot with CHOIR clusters labeled

Generate a dimensionality reduction embedding plot colored (by default) according to the clusters identified by CHOIR. Additionally permits plotting mean prediction accuracy scores for pairs of clusters as an overlay.

Usage

plotCHOIR(
  object,
  key = "CHOIR",
  reduction = NULL,
  group_by = NULL,
  cells = NULL,
  highlight_cells = NULL,
  label = TRUE,
  legend = TRUE,
  accuracy_scores = FALSE,
  plot_nearest = TRUE,
  ...
)

Arguments

object

An object of class Seurat, SingleCellExperiment, or ArchRProject that has undergone CHOIR clustering. For multi-omic data, we recommend using ArchRProject objects.

key

The name under which CHOIR-related data for this run is stored in the object. Defaults to “CHOIR”.

reduction

A character vector indicating which CHOIR subtree dimensionality reduction to plot (e.g., 'P0_reduction_UMAP', 'P1_reduction_UMAP'). Default = NULL will plot the UMAP of the 'P0_reduction_UMAP' generated by CHOIR.

group_by

A character string indicating which metadata column to color cells by. Default = NULL will color cells according to the clusters identified by CHOIR.

cells

Vector of cell names to plot. Default is all cells.

highlight_cells

Vector of cell names to highlight. Default = NULL.

label

A Boolean value indicating whether to label groups. Defaults to TRUE.

legend

A Boolean value indicating whether to include a legend. Defaults to TRUE.

accuracy_scores

A Boolean value indicating whether to overlay pairwise cluster comparison accuracy scores onto UMAP. Default = TRUE.

plot_nearest

A Boolean value indicating whether to only plot the accuracy score for each cluster's nearest neighboring cluster. Only relevant if accuracy_scores is TRUE. Defaults to TRUE.

...

Extra parameters passed to Seurat::DimPlot().

Value

A ggplot object.